+++ app/src/androidTest/java/com/univalle/parcial/parcial/activities/ActivityLogin.java
@@ -1,4 +1,4 @@
-package com.univalle.parcial.parcial;
+package com.univalle.parcial.parcial.activities;
 
 import android.app.AlertDialog;
 import android.content.Intent;
@@ -10,6 +10,7 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.univalle.parcial.parcial.R;
 import com.univalle.parcial.parcial.conexion.ConexionBD;
 import com.univalle.parcial.parcial.conexion.UsuarioBD;
 import com.univalle.parcial.parcial.modelo.Usuario;
