+++ app/src/androidTest/java/com/univalle/parcial/parcial/fragments/ProductoFragment.java
@@ -1,4 +1,4 @@
-package com.univalle.parcial.parcial;
+package com.univalle.parcial.parcial.fragments;
 
 import android.content.Context;
 import android.net.Uri;
@@ -8,16 +8,18 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.univalle.parcial.parcial.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link Minutos.OnFragmentInteractionListener} interface
+ * {@link ProductoFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link Minutos#newInstance} factory method to
+ * Use the {@link ProductoFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Minutos extends Fragment {
+public class ProductoFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -29,7 +31,7 @@ public class Minutos extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public Minutos() {
+    public ProductoFragment() {
         // Required empty public constructor
     }
 
@@ -39,11 +41,11 @@ public class Minutos extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Minutos.
+     * @return A new instance of fragment ProductoFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static Minutos newInstance(String param1, String param2) {
-        Minutos fragment = new Minutos();
+    public static ProductoFragment newInstance(String param1, String param2) {
+        ProductoFragment fragment = new ProductoFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -64,7 +66,7 @@ public class Minutos extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_minutos, container, false);
+        return inflater.inflate(R.layout.fragment_producto, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
