+++ app/src/main/res/layout/fragment_registro__producto.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Registro_Producto">
+    tools:context=".fragments.Registro_Producto">
 
     <!-- TODO: Update blank fragment layout -->
